+++ library/data-model/src/data_storage/authDB/init.ts
@@ -4,7 +4,11 @@ import {AuthDatabaseSecurityDocument} from './security';
 
 // currently no config for the auth DB
 export type AuthDbInitialisationConfig = {};
-export function initAuthDB({}: AuthDbInitialisationConfig): InitialisationContent {
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export function initAuthDB(
+  config: AuthDbInitialisationConfig
+): InitialisationContent {
   return {
     designDocuments: [viewsDocument, permissionDocument],
     securityDocument: AuthDatabaseSecurityDocument,
