+++ library/data-model/src/data_storage/dataDB/design.ts
@@ -1,7 +1,7 @@
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-nocheck
 
-import { convertToCouchDBString } from "../utils";
+import {convertToCouchDBString} from '../utils';
 
 /**
  * Design document for filtering attachments
@@ -24,19 +24,21 @@ export const attachmentFilterDocument = {
  */
 export const permissionsDocument = {
   _id: '_design/permissions',
-  validate_doc_update: convertToCouchDBString((newDoc, oldDoc, userCtx, _secObj) => {
-    if (userCtx === null || userCtx === undefined) {
-      throw {
-        unauthorized: 'You must be logged in. No token given.',
-      };
+  validate_doc_update: convertToCouchDBString(
+    (newDoc, oldDoc, userCtx, _secObj) => {
+      if (userCtx === null || userCtx === undefined) {
+        throw {
+          unauthorized: 'You must be logged in. No token given.',
+        };
+      }
+      if (userCtx.name === null || userCtx.name === undefined) {
+        throw {
+          unauthorized: 'You must be logged in. No username given.',
+        };
+      }
+      return;
     }
-    if (userCtx.name === null || userCtx.name === undefined) {
-      throw {
-        unauthorized: 'You must be logged in. No username given.',
-      };
-    }
-    return;
-  }),
+  ),
 };
 
 /**
@@ -47,8 +49,7 @@ export const indexDocument = {
   views: {
     record: {
       map: convertToCouchDBString(doc => {
-        if (doc.record_format_version === 1)
-          emit(doc._id, 1);
+        if (doc.record_format_version === 1) emit(doc._id, 1);
       }),
     },
     recordRevisions: {
@@ -59,26 +60,29 @@ export const indexDocument = {
             const created = doc.created;
             const created_by = doc.created_by;
             const type = doc.type;
-            emit(doc._id, {_id: doc.heads[0], conflict, created, created_by, type});
+            emit(doc._id, {
+              _id: doc.heads[0],
+              conflict,
+              created,
+              created_by,
+              type,
+            });
           }
       }),
     },
     revision: {
       map: convertToCouchDBString(doc => {
-        if (doc.revision_format_version === 1)  
-          emit(doc._id, 1);
+        if (doc.revision_format_version === 1) emit(doc._id, 1);
       }),
     },
     avp: {
       map: convertToCouchDBString(doc => {
-        if (doc.avp_format_version === 1)
-          emit(doc._id, 1);
+        if (doc.avp_format_version === 1) emit(doc._id, 1);
       }),
     },
     recordCount: {
       map: convertToCouchDBString(doc => {
-        if (doc.record_format_version === 1)
-          emit(doc._id, 1);
+        if (doc.record_format_version === 1) emit(doc._id, 1);
       }),
       reduce: '_count',
     },
@@ -91,5 +95,5 @@ export const indexDocument = {
 export const dataDbDesignDocuments = {
   attachmentFilterDocument,
   permissionsDocument,
-  indexDocument
+  indexDocument,
 };
