+++ library/data-model/src/data_storage/directoryDB/design.ts
@@ -1,8 +1,6 @@
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-nocheck
 
-import {convertToCouchDBString} from '../utils';
-
 /**
  * Exports all design documents for the directory database
  * (none)
