+++ library/data-model/src/data_storage/metadataDB/design.ts
@@ -1,8 +1,6 @@
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-nocheck
 
-import {convertToCouchDBString} from '../utils';
-
 /**
  * Exports all design documents for the metadata database
  * (None)
