+++ library/data-model/src/data_storage/peopleDB/init.ts
@@ -1,8 +1,11 @@
-import {InitialisationContent} from '../utils';
+import {InitialisationContent} from '../';
 import {PeopleDBSecurityDocument} from './security';
 
 export type PeopleDBInitialisationConfig = {};
-export function initPeopleDB({}: PeopleDBInitialisationConfig): InitialisationContent {
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export function initPeopleDB(
+  config: PeopleDBInitialisationConfig
+): InitialisationContent {
   return {
     designDocuments: [],
     securityDocument: PeopleDBSecurityDocument,
