+++ library/data-model/src/data_storage/projectsDB/init.ts
@@ -2,7 +2,10 @@ import {InitialisationContent} from '../utils';
 import {ProjectsDBSecurityDocument} from './security';
 
 export type ProjectsDBInitialisationConfig = {};
-export function initProjectsDB({}: ProjectsDBInitialisationConfig): InitialisationContent {
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export function initProjectsDB(
+  config: ProjectsDBInitialisationConfig
+): InitialisationContent {
   return {
     designDocuments: [],
     securityDocument: ProjectsDBSecurityDocument,
