+++ library/data-model/src/data_storage/templatesDB/index.ts
@@ -1,4 +1,4 @@
 // export * from './types';
 export * from './security';
 export * from './design';
-export * from './init'
+export * from './init';
