+++ library/data-model/src/data_storage/templatesDB/init.ts
@@ -2,7 +2,10 @@ import {InitialisationContent} from '../utils';
 import {TemplatesDBSecurityDocument} from './security';
 
 export type TemplatesDBInitialisationConfig = {};
-export function initTemplatesDB({}: TemplatesDBInitialisationConfig): InitialisationContent {
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export function initTemplatesDB(
+  config: TemplatesDBInitialisationConfig
+): InitialisationContent {
   return {
     designDocuments: [],
     securityDocument: TemplatesDBSecurityDocument,
