+++ library/data-model/src/datamodel/index.ts
@@ -1,0 +1,4 @@
+// Files in this folder
+export * from './core';
+export * from './database';
+export * from './typesystem';
